Make venv target update mtime on venv/
Changing requirements.txt and installing updates into venv/ doesn't seem
to change the mtime on the directory (at least on my mac), causing the
venv/ dependency for all targets to cause noisy no-op pip installations
all the time.

Explicity touching the directory fixes that, although I wonder if this
is the idiomatic way to do so.